v1.2.7
Had to rewrite a ton of stuff to get this working 100%, for some reason still it is trying to run the execute script more times than it supposed to, it should check if the availabel ram is above the scripts required and then run but it seems to keep running the loop. but it works a charm.